// src/types.ts
export interface PetInfo {
  name: string;
  species: string;
  age: string;
  favoriteFood: string;
  accent: string;
  extraInfo: string;
  photo: string;
}

export interface Theme {
  id: string;
  name: string;
  bgColor: string;
  textColor: string;
  accentColor: string;
  secondaryColor: string;
}